requirements.txt
chromadb>=0.5.0
google-generativeai>=0.8.0
rank-bm25>=0.2.2
streamlit>=1.32.0
numpy>=1.26.0
scikit-learn>=1.4.0
